fixed all todos
