+++ README.md
@@ -223,10 +223,3 @@ import { mount } from 'enzyme';
 # Debugging
 1. By default log level is `info`
 2. Set the log level to `verbose` by importing and setting `configureLogLevel('verbose')`
-
-# Testing: TODO: remove later
-## locally:
-1. bump the version in package.json
-1. `npm run build`
-1. `npm pack`
-1. install in your host project with `npm install <path_to_tar_file>`
